fix: 8th audit (v1.9.0+) — 23 of 88 findings fixed (P1 + P2) (#715)
* fix: 8th audit (v1.9.0+) — 23 of 88 findings fixed (P1 + P2)

8th full audit: 88 findings across 14 categories in 3 batches.

P1 fixes (12):
- RB-34: dim=0 guard in HNSW build (prevents chunks_exact(0) panic)
- RB-29: empty chunk name guard + zero-vector warning in embedder
- DS-33: delete_phantom_chunks temp table (SQLite 999-param limit)
- RM-37: chunked encoding size cap via Read::take()
- PB-32/SEC-25: dunce::canonicalize on CQS_ONNX_DIR
- DOC-38/42/44/47, AD-44/47: 9 stale E5-base/768 → BGE-large/1024

P2 fixes (11 of 15):
- AC-25: BFS inner-loop node cap guard (all 4 functions)
- AC-28: full_cosine_similarity f64 accumulators at 1024-dim
- EH-40: resume() single hash call (TOCTOU fix)
- EH-43: submit_fresh propagates set_pending failure
- RM-39: drop matrix/sims early in contrastive neighbors (~400MB saved)
- SEC-28: custom embedding repo validation
- TC-43: 8 tests for SEC-20 path traversal + SEC-28 repo validation
- PERF-44: notes loaded once per search (was twice)
- PERF-46: Arc<Vec<NoteSummary>> cache (was deep-clone per search)
- DS-35: watch loads existing HNSW on restart (orphan prevention)
- RB-31: grammar() → try_grammar() at all 9 parser call sites

Also: audit skill updated with prompt-generation + review steps,
openclaw contributions tracker, future training experiments in roadmap.


* fix: clippy warnings + fmt from P2 fixes


* fix: P3 audit fixes — 35 of 39 findings (docs, tracing, errors, tests, API)

P3 fixes applied:
- DOC-39/40/41/43, AD-45/46, RM-36: 10 stale E5-base/768 doc comments
- DOC-45: CONTRIBUTING.md add cqs-verify skill
- DOC-46: CQS_ONNX_DIR documented in README + SECURITY.md
- AD-50: VectorIndex::dim() trait method added
- AD-53: ModelInfo.dimensions u32→usize (removes casts everywhere)
- AD-54: Embedding::new vs try_new doc guidance
- OB-28..32: tracing spans for detect_provider, create_session,
  parse_unified_diff, find_changed_functions, audit state, embeddings batch
- EH-41/42/44: warn on silently swallowed DB/HTTP errors
- CQ-35: dead check_model_version → #[cfg(test)]
- RB-30: fragile checkpoint_sha unwrap → is_some_and
- AC-29/30: documented edge case behavior
- DS-36: HNSW backup rename failure now logged
- PERF-41: use make_placeholders helper
- RM-35: clear_session clears query cache
- EX-36: doc_format validation test for all 51 languages
- EX-37/38: download size comment, CAGRA_THRESHOLD env var override
- PB-35: onnx_path forward-slash contract documented
- SEC-26: api_base credential redaction in logs
- SEC-27: git_show path colon validation
- TC-42/44/49: 9 new tests (phantom chunks, SearchFilter, validate_f32)

Also: CLAUDE.md workflow examples, ROADMAP agent adoption updates,
pre-edit hook, v9-200k-1.5ep results, research log update.

Skipped: AD-48 (low benefit), PERF-42 (not stale), RB-32 (trait change),
EX-36 full enum (test approach instead).


* fix: P4 audit fixes — 18 of 22 findings (docs, API, tests, security, perf)

Trivial (6): AC-26/27/31 edge case docs, PB-33 display→to_string_lossy,
RM-38 git_log 100K warning, SEC-29 webhelp symlink check.

Easy (6): AD-52 ModelInfo→embedder::models, EX-39 hyde_max_tokens in
LlmConfig, TC-48 NaN clamp fix+test, TC-45 CQS_ONNX_DIR tests,
AD-51/EX-35 doc improvements.

Medium (5): AD-49 shared OutputArgs/TextJsonArgs structs (replaced
per-command boilerplate), CQ-39 ParserStageContext struct, DS-34 notes
lock, DS-37 phantom delete error propagation, RB-33 file lock on
rewrite_file.

Tests (1): TC-46 batch error-path tests via MockBatchProvider.

Skipped: CQ-37 (nl/generation.rs split too tightly coupled),
AD-51 full merge (40+ call sites, cosmetic).

1511 tests, 0 failures, 0 clippy warnings.


* fix: add llm_hyde_max_tokens to test Config constructors (CI fix)


---------

+++ src/cli/commands/reference.rs
@@ -124,7 +124,7 @@ fn cmd_ref_add(cli: &Cli, name: &str, source: &std::path::Path, weight: f32) -> 
             .with_context(|| format!("Failed to open reference store at {}", db_path.display()))?,
     );
     let mc = cli.model_config();
-    store.init(&ModelInfo::new(&mc.repo, mc.dim as u32))?;
+    store.init(&ModelInfo::new(&mc.repo, mc.dim))?;
     let stats = run_index_pipeline(
         &source,
         files,
